Factor common declaration and comment handling functionality to base code generator.
